Long to Bigdecimal AS-7083

+++ src/main/java/com/asg/shipchandling/salesinvoice/dto/request/CreateSalesInvoiceDtlRequest.java
@@ -14,10 +14,10 @@ import java.math.BigDecimal;
 public class CreateSalesInvoiceDtlRequest {
     private Long stockPoid;
     private Long stockUnitPoid;
-    private Long quantity;
+    private BigDecimal quantity;
     private BigDecimal price;
-    private Long discount;
-    private Long baseAmt;
+    private BigDecimal discount;
+    private BigDecimal baseAmt;
     private Long taxPoid;
     private String costCenterPoid;
     private String remarks;
